# docs/swagger.yaml
basePath: /
definitions:
  entity.UserSignIn:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  entity.UserSignUp:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - name
    - password
    - username
    type: object
  server.errorResponse:
    properties:
      message:
        type: string
    type: object
  server.refreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  server.tokens:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for Authentication
  title: Authentication API
  version: "1.0"
paths:
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: refresh token
      operationId: refresh-token
      parameters:
      - description: refresh token
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/server.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.tokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/server.errorResponse'
      summary: RefreshToken
      tags:
      - auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: login
      operationId: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.UserSignIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.tokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/server.errorResponse'
      summary: SignIn
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: create account
      operationId: create-account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.UserSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.tokens'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/server.errorResponse'
      summary: SignUp
      tags:
      - auth
swagger: "2.0"